Refactored the client validator to auth
