+++ ms-common-auth/src/main/java/net/trajano/ms/auth/jsonclientvalidator/ClientInfo.java
@@ -1,4 +1,4 @@
-package net.trajano.ms.example.authz.sample;
+package net.trajano.ms.auth.jsonclientvalidator;
 
 import java.net.URI;
 import java.util.HashSet;
