+++ ms-common-auth/src/main/java/net/trajano/ms/auth/jsonclientvalidator/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * An implementation of the client validator that uses a JSON file source.
+ *
+ * @author Archimedes Trajano
+ */
+package net.trajano.ms.auth.jsonclientvalidator;
