+++ ms-common-auth/src/main/java/net/trajano/ms/auth/spi/ClientValidator.java
@@ -1,6 +1,9 @@
 package net.trajano.ms.auth.spi;
 
 import java.net.URI;
+import java.text.ParseException;
+
+import net.trajano.ms.auth.util.HttpAuthorizationHeaders;
 
 /**
  * This is used to validate the client ID and secret.
@@ -19,6 +22,13 @@ public interface ClientValidator {
     default URI getJwksUri(final String clientId) {
 
         return null;
+    }
+
+    default boolean isValid(final String grantType,
+        final String authorization) throws ParseException {
+
+        final String[] authInfo = HttpAuthorizationHeaders.parseBasicAuthorization(authorization);
+        return isValid(grantType, authInfo[0], authInfo[1]);
     }
 
     boolean isValid(String grantType,
